[Http] Move misfiled RouteFactory from Cli to Http namespace.
