+++ src/Valkyrja/Http/Routing/Collector/AttributeRouteCollector.php
@@ -18,7 +18,6 @@ use ReflectionException;
 use ReflectionMethod;
 use Valkyrja\Attribute\Collector\Collector;
 use Valkyrja\Attribute\Collector\Contract\CollectorContract;
-use Valkyrja\Cli\Routing\Factory\RouteFactory;
 use Valkyrja\Http\Middleware\Contract\ResponseSentMiddlewareContract;
 use Valkyrja\Http\Middleware\Contract\RouteDispatchedMiddlewareContract;
 use Valkyrja\Http\Middleware\Contract\RouteMatchedMiddlewareContract;
@@ -40,6 +39,7 @@ use Valkyrja\Http\Routing\Data\Contract\RouteContract;
 use Valkyrja\Http\Routing\Data\DynamicRoute;
 use Valkyrja\Http\Routing\Data\Parameter as DataParameter;
 use Valkyrja\Http\Routing\Data\Route;
+use Valkyrja\Http\Routing\Factory\RouteFactory;
 use Valkyrja\Http\Routing\Processor\Contract\ProcessorContract;
 use Valkyrja\Http\Routing\Processor\Processor;
 use Valkyrja\Http\Struct\Request\Contract\RequestStructContract;
