+++ src/Valkyrja/Http/Routing/Factory/RouteFactory.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Cli\Routing\Factory;
+namespace Valkyrja\Http\Routing\Factory;
 
 use Valkyrja\Http\Routing\Data\Contract\DynamicRouteContract;
 use Valkyrja\Http\Routing\Data\Contract\RouteContract;
