+++ tests/Tests/Unit/Http/Routing/Factory/RouteFactoryTest.php
@@ -11,13 +11,13 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Tests\Unit\Cli\Routing\Factory;
+namespace Valkyrja\Tests\Unit\Http\Routing\Factory;
 
-use Valkyrja\Cli\Routing\Factory\RouteFactory;
 use Valkyrja\Http\Message\Enum\RequestMethod;
 use Valkyrja\Http\Routing\Data\DynamicRoute;
 use Valkyrja\Http\Routing\Data\Parameter;
 use Valkyrja\Http\Routing\Data\Route;
+use Valkyrja\Http\Routing\Factory\RouteFactory;
 use Valkyrja\Tests\Fixtures\Http\Struct\IndexedJsonRequestStructEnum;
 use Valkyrja\Tests\Fixtures\Http\Struct\ResponseStructEnum;
 use Valkyrja\Tests\Unit\Abstract\TestCase;
